Make it per-namespace
